requirements.txt
flask==3.1.0
python-dotenv==1.1.0
requests==2.32.3
gunicorn==23.0.0
line-bot-sdk==3.14.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.37
